requirements.txt
Flask==2.2.5
opencv-python==4.8.1
numpy==1.26.4
Pillow==10.1.0
tensorflow==2.14.0 
